u c

+++ 3_ft_printf/srcs/print_unsigned_int.c
@@ -6,7 +6,7 @@
 /*   By: junhpark <junhpark@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/25 17:20:01 by junhpark          #+#    #+#             */
-/*   Updated: 2020/07/01 16:40:16 by junhpark         ###   ########.fr       */
+/*   Updated: 2020/07/01 17:12:40 by junhpark         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ int			ft_unsigned_int(char *data, va_list ap, t_flag *data_flag)
 			flag_width = va_arg(ap, int);
 		idx++;
 	}
-	num = (unsigned long long)va_arg(ap, unsigned int);
+	num = va_arg(ap, unsigned int);
 	data_flag->minus_flag = FALSE;
 	input_string = ft_itoa_u(num);
 	make_int_flag(data_flag, data, flag_width, input_string);
